+++ lua/smaug/init.lua
@@ -9,11 +9,13 @@
 
 local Series     = require("smaug.core.series")
 local BoolSeries = require("smaug.core.boolseries")
+local DataSet    = require("smaug.core.dataset")
 
 local smaug = {
-    _VERSION = "0.2.0-dev",   -- Fase 2 em andamento
+    _VERSION = "0.3.0-dev",   -- Fase 3 em andamento
     Series     = Series,
     BoolSeries = BoolSeries,
+    DataSet    = DataSet,
 }
 
 -- Açúcar: smaug.float64(...) == smaug.Series.float64(...)
@@ -21,5 +23,6 @@ smaug.float64    = Series.float64
 smaug.int64      = Series.int64
 smaug.from_table = Series.from_table
 smaug.NA = Series.NA
+smaug.dataset = DataSet.from_columns   -- smaug.dataset({{"a", {1,2}}, {"b", {3,4}}})
 
 return smaug